requirements.txt
python-frontmatter==1.1.0
jinja2==3.1.3
fastapi==0.110.0
uvicorn==0.27.1
GitPython==3.1.43
markdown==3.10.2